Season length is now defined as value 0 - 1.0

+++ src/config.rs
@@ -34,7 +34,7 @@ pub struct GameConfig {
     pub person_ttl: ConfigValue<u32>,
     pub food_for_baby: ConfigValue<u32>,
     pub year_length: ConfigValue<u32>,
-    pub growing_season_length: ConfigValue<u32>,
+    pub growing_season_length: ConfigValue<f32>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Component)]
